Move table properties classes to own namespace.
